move config tests to new file
